gtfs: fix incorrect interpretation of stop_times.txt offsets
These were interpreted as "hours from midnight of date", but that is wrong.
Sometimes it can be an offset from 01:00, or 23:00 of the previous day.

Upside is that these allow for simple ariphmetics between them, without
having to worry about DST on every day of the trip with 24h+ offsets.

GTFS doc quote:
 The time is measured from "noon minus 12h"
  (effectively midnight, except for days on which daylight
  savings time changes occur) at the beginning of the service date.

https://developers.google.com/transit/gtfs/reference/stop_times-file